login + register with firebase
