+++ teg_auto/pubspec.yaml
@@ -31,6 +31,7 @@ environment:
 dependencies:
   cloud_firestore: ^4.2.0
   cupertino_icons: ^1.0.2
+  firebase_auth: ^4.2.1
   firebase_core: ^2.4.0
   flutter:
     sdk: flutter
